sécurité : verrou d'accès à la carte SD, et suites d'audit
Deux audits lancés sur le code du jour. Le reviewer n'a trouvé aucun
bloquant ; l'auditeur sécurité, une course réelle, déclenchable depuis
l'hôte.

CRITIQUE — sd_probe() et les callbacks MSC tournent dans deux tâches
distinctes, sur un P4 bicœur, sans exclusion mutuelle. La console est
accessible à l'hôte par la CDC de l'USB-Serial-JTAG : un « sd probe »
pendant un transfert libérait s_card sous les pieds de la tâche USB, et
détruisait le mutex interne du pilote SDMMC que celle-ci était en train de
prendre. Usage après libération sur le tas et sémaphore détruit en cours
d'usage, tous deux pilotables par du logiciel hôte seul.

Corrigé par un mutex alloué statiquement, pris par sd_probe, sd_read_sectors
et sd_write_sectors. Il couvre l'appel SDMMC entier et pas seulement le
contrôle de bornes : la fenêtre s'ouvre pendant que le pilote rend la main.
sd_card_init() le crée avant que l'USB ne démarre.

Le deinit de l'hôte SDMMC devient conditionnel : au premier sondage il
n'avait jamais été initialisé, donc l'appel inconditionnel échouait
légitimement à chaque démarrage — ce qui rendait impossible de journaliser
ses vrais échecs. Ils le sont maintenant, ici comme pour le LDO et le PHY
USB.

Aussi, tous confirmés puis corrigés : un LBA hors bornes était signalé
MEDIUM ERROR au lieu d'ILLEGAL REQUEST, ce qui pouvait faire passer une
carte saine pour morte auprès de l'OS en quelques CDB ; l'extension de
signe sur les chaînes de descripteurs USB, latente tant qu'elles sont
ASCII ; link_proto_parse_status indexait des offsets fixes sans recevoir
de longueur, refermé avant que son premier appelant n'existe.

Du reviewer : commentaires renvoyant à main/board.h, supprimé ce matin
même ; en-tête de fast.sh affirmant qu'il n'y a pas de tests alors qu'il
en lance ; BOARD_LINK_AVAILABLE non gardé — une carte qui l'oubliait
compilait en silence avec le lien désactivé, le préprocesseur traitant une
macro inconnue comme 0. Un #error ferme ça, et j'ai vérifié qu'il mord.
Avis TDD de check.sh armé, il a désormais quelque chose à surveiller.

Consigné dans docs/HARDWARE.md, sans code à changer : l'hôte peut forcer
le mode download par les lignes de contrôle de la CDC (TRM chap. 53, table
53.3-2, p. 2715 — observé aussi ce matin par accident). Sans conséquence
aujourd'hui, mais ça invalide l'hypothèse « clés en clair en NVS parce que
l'extraction demande un accès physique ». À trancher avant la première clé.

Vérifié sur le kit : 317 assertions hôte, les deux builds, et aucune
régression — 4650 secteurs en DMA direct, 0 rebond, carte détectée par le
chemin d'alimentation externe.

+++ main/usb/usb_device.c
@@ -169,7 +169,11 @@ uint16_t const *tud_descriptor_string_cb(uint8_t index, uint16_t langid)
             len = max;   /* tronqué plutôt que débordé */
         }
         for (size_t i = 0; i < len; i++) {
-            desc[i + 1] = (uint16_t)str[i];   /* ASCII vers UTF-16LE */
+            /* Le double cast n'est pas décoratif : `char` est signé sur RISC-V,
+             * et un octet ≥ 0x80 deviendrait 0xFFxx en UTF-16 par extension de
+             * signe. Inoffensif tant que les chaînes sont ASCII, silencieux le
+             * jour où l'une sera accentuée. */
+            desc[i + 1] = (uint16_t)(uint8_t)str[i];   /* vers UTF-16LE */
         }
         count = (uint8_t)len;
     }
